IS-3 Move Google and PlanningCenter clients
